Test.Hspec.Extra

Description

Helper functions for testing.

Documentation

aroundAll :: forall a. HasCallStack => (ActionWith a -> IO ()) -> SpecWith a -> Spec

Run a bracket resource acquisition function around all the specs. The resource is allocated just before the first test case and released immediately after the last test case.

Each test is given the resource as a function parameter.

it :: HasCallStack => String -> ActionWith ctx -> SpecWith ctx

A drop-in replacement for it that'll automatically retry a scenario once if it fails, to cope with potentially flaky tests, if the environment variable TESTS_RETRY_FAILED is set.

It also has a timeout of 10 minutes.

itWithCustomTimeout

Arguments

:: HasCallStack

=> Int
Timeout in seconds.

-> String

-> ActionWith ctx

-> SpecWith ctx

Like it but with a custom timeout, testing of the function possible.

parallel :: SpecWith a -> SpecWith a

Like Hspec's parallel, except on Windows.

counterexample :: (MonadIO m, MonadUnliftIO m, HasCallStack) => String -> m a -> m a

Can be used to add context to a HUnitFailure.

>>> counterexample (show response) (0 `shouldBe` 3)
>>> (Status {statusCode = 200, statusMessage = "OK"},Right [])
>>> expected: 3
>>> but got: 0

appendFailureReason :: String -> HUnitFailure -> HUnitFailure

Custom test suite runner

type HspecWrapper a = IO Summary -> IO a

An IO action that runs around hspecWith.

hspecMain :: Spec -> IO ()

Main function for running a test suite using getDefaultConfig.

hspecMain' :: IO (HspecWrapper a, Config) -> Spec -> IO a

Main function for running a test suite. Like hspec, except it allows for a custom action to modify the environment and configuration before passing control over to Hspec.

getDefaultConfig :: IO (HspecWrapper (), Config)

Our custom Hspec wrapper. It adds the --env option for setting environment variables, and prints the tests which took the longest time after finishing the test suite.

Internals

configWithExecutionTimes :: Config -> Config

Add execution timing information to test output.

setEnvParser :: ParserInfo ([(String, String)], [String])

A CLI arguments parser which handles setting environment variables.