Test.Hspec.Goldens

Description

Small module for writing text-based golden tests.

In contrast to `hspec-golden-aeson` this module works with manually provided values instead of QuickCheck, and with text instead of JSON.

If the tests are run with the OVERWRITE_GOLDENS environment variable set, they will all fail, and write the current values to disk.

If some goldens are os-dependent (like some optparse-applicative help text using getProgName), you can create a windows specific variant by appending ".win" to the golden file name.

This seems to work for now with just two windows specific goldens, but if there are more, it might be more convenient if windows goldens always are separate, as OVERWRITE_GOLDENS will never currently write the .win version.

Documentation

data Settings

Constructors

Settings

Fields

goldenDirectory :: FilePath

postProcess :: Text -> Text

textGolden

Arguments

:: Settings

-> String
Filename for the test

-> Text
Value to compare with the golden file

-> Expectation