Test.QuickCheck.Extra

Description

Extra helper functions for QuickCheck

Generation

genFunction :: (a -> Gen b -> Gen b) -> Gen b -> Gen (a -> b)

Generates a function.

This is based on the implementation of Arbitrary for 'a -> b'.

genMapWith :: Ord k => Gen k -> Gen v -> Gen (Map k v)

Generates a Map with the given key and value generation functions.

genSized2 :: Gen a -> Gen b -> Gen (a, b)

Generates a 2-tuple whose range depends linearly on the size parameter.

genSized2With :: (a -> b -> c) -> Gen a -> Gen b -> Gen c

Similar to genSized2, but with a custom constructor.

reasonablySized :: Gen a -> Gen a

Resize a generator to grow with the size parameter, but remains reasonably sized. That is handy when testing on data-structures that can be arbitrarily large and, when large entities don't really bring any value to the test itself.

It uses a square root function which makes the size parameter grows quadratically slower than normal. That is,

Normal Size ----------- 0 1 10 100 1000	Reasonable Size ---------------- 0 1 3 10 31

Shrinking

liftShrinker :: (a -> [a]) -> (I -.-> []) a

Lift a shrinking function into a type more convenient for use with Generics.SOP.

Examples:

shrinkBool :: Bool -> [Bool]
shrinkChar :: Char -> [Char]

data BoolChar = BoolChar Bool Char

instance Generic BoolChar

shrinkBoolChar :: BoolChar -> [BoolChar]
shrinkBoolChar =
    groundRobinShrink (  liftShrinker shrinkBool
                      :* liftShrinker shrinkChar
                      :* Nil
                      )

shrinkInterleaved :: (a, a -> [a]) -> (b, b -> [b]) -> [(a, b)]

Shrink the given pair in interleaved fashion.

Successive shrinks of the left and right hand sides are interleaved in the resulting sequence, to avoid biasing either side.

shrinkMapWith :: Ord k => (k -> [k]) -> (v -> [v]) -> Map k v -> [Map k v]

Shrinks a Map with the given key and value shrinking functions.

groundRobinShrink

Arguments

:: (Generic a, Code a ~ '[xs])

=> NP (I -.-> []) xs
and a list of shrinking functions, one for each argument of the product

-> a
and a value of that type

-> [a]
provide a list of shrunk values.

Given a list of shrinkers for each element of a product type (NOTE: this function is not defined for sum types), and a value of that product type, shrink the value using a round-robin algorithm.

Examples:

shrinkBoolChar :: BoolChar -> [BoolChar]
shrinkBoolChar =
    groundRobinShrink (  liftShrinker shrinkBool
                      :* liftShrinker shrinkChar
                      :* Nil
                      )

groundRobinShrink'

Arguments

:: (Generic a, Code a ~ '[xs], All Arbitrary xs)

=> a
Given such a type

-> [a]
return a shrunk list of that product type, using the round-robin algorithm.

Same as groundRobinShrink, but use the available shrinkers via Arbitrary instance.

Examples:

shrinkBoolChar :: BoolChar -> [BoolChar]
shrinkBoolChar = groundRobinShrink'

genericRoundRobinShrink :: (Generic a, GFrom a, GTo a, GCode a ~ '[xs]) => NP (I -.-> []) xs -> a -> [a]

This function exists to provide a GHC.Generics version of groundRobinShrink, so that users of this code don't have to derive an instance of Generics.SOP.Generic; an instance of GHC.Generics.Generic will do.

genericRoundRobinShrink' :: (Generic a, GFrom a, GTo a, GCode a ~ '[xs], All Arbitrary xs) => a -> [a]

Same as genericRoundRobinShrink but uses available Arbitrary instance for shrinking.

(<@>) :: (a -> b) -> a -> b infixl 6

(<:>) :: (x -> [x]) -> NP (I -.-> []) xs -> NP (I -.-> []) (x ': xs) infixr 7

Evaluating shrinkers

genShrinkSequence :: forall a. (a -> [a]) -> a -> Gen [a]

Generates a random sequence of progressively shrunken values from a given starting value and shrinking function.

Each successive element in the sequence is selected at random from the result of applying the shrinking function to the preceding element.

The given starting value is not included in the sequence, by default.

Examples:

>>> generate (genShrinkSequence shrink (100 :: Int))
[94,83,82,72,70,66,33,32,16,0]

>>> generate (genShrinkSequence shrink "Cardano")
["Caraano","aaraano","aaraaao","aaro","aarb","aaab","aab","aa",""]

The resulting sequence will be empty if (and only if) applying the shrinking function to the starting value yields the empty list:

>>> generate (genShrinkSequence (const []) "Cardano")
[]

If the resulting sequence is non-empty, then applying the shrinking function to the terminal element will yield the empty list:

>>> shrink . last <$> generate (genShrinkSequence shrink (100 :: Int))
[]

shrinkSpace :: forall a. Ord a => (a -> [a]) -> a -> Set a

Computes the shrink space of a given shrinking function for a given starting value.

This function returns the set of all values that are transitively reachable through repeated applications of the given shrinking function to the given starting value.

By default, the given starting value is not included in the result.

Examples:

>>> shrinkSpace shrink "abc"
["","a","aa","aaa","aab","aac","ab","aba","abb","ac","b","ba","bb","bc","c"]

>>> shrinkSpace shrink (8 :: Int)
[0,1,2,3,4,5,6,7]

>>> shrinkSpace shrink (2 :: Int, 2 :: Int)
[(0,0),(0,1),(0,2),(1,0),(1,1),(1,2),(2,0),(2,1)]

Caution:

Depending on the particular choice of shrinking function and starting value, the shrink space can grow very quickly. Therefore, this function should be used with caution to avoid non-termination within test cases. If in doubt, use the within modifier provided by QuickCheck to ensure that your test case terminates within a fixed time limit.

This function can be used to test that a given shrinking function always generates values that satisfy a given condition. For example:

prop_shrinkApple_valid :: Apple -> Property
prop_shrinkApple_valid apple =
    within twoSeconds $
    all isValidApple (shrinkSpace shrinkApple apple)
  where
    twoSeconds = 2_000_000

shrinkWhile :: (a -> Bool) -> (a -> [a]) -> a -> Maybe a

Repeatedly applies a shrinking function to a value while a condition holds.

This function can be used to predict the final value that QuickCheck will produce when searching for a minimal counterexample.

Example:

>>> isCounterexample a = (a > 0) && (a `mod` 8 == 0)
>>> shrinkWhile isCounterexample shrinkIntegral 1024
Just 8

Provided that the given starting value satisfies the condition, and provided that at least one shrunken value satisfies the condition, this function will terminate with the smallest shrunken value that cannot be shrunk further with the given shrinking function.

This function returns Nothing if the given starting value does not satisfy the condition, or if none of the shrunken values satisfy the condition.

The final result is evaluated eagerly. If you suspect that a given shrinking sequence does not terminate, then you may wish to consider evaluating a finite prefix of shrinkWhileSteps instead.

shrinkWhileSteps :: forall a. (a -> Bool) -> (a -> [a]) -> a -> [a]

Repeatedly applies a shrinking function to a value while a condition holds, returning all the intermediate shrinking steps.

This function can be used to predict the sequence of intermediate values that QuickCheck will produce when searching for a minimal counterexample.

Example:

>>> isCounterexample = (>= 100)
>>> shrinkWhileSteps isCounterexample shrinkIntegral 1024
[512,256,128,112,105,102,101,100]

Provided that the given starting value satisfies the condition, and provided that at least one shrunken value satisfies the condition, this function will produce a non-empty list of all intermediate shrinking steps, ordered from largest to smallest.

The list is evaluated lazily from largest to smallest. If you suspect that a given shrinking sequence does not terminate, then you may wish to consider evaluating a finite prefix of the list.

This function returns the empty list if the given starting value does not satisfy the condition, or if none of the shrunken values satisfy the condition.

Partitioning lists

partitionList

Arguments

:: (Int, Int)
The minimum and maximum length parameters.

-> [a]
The list to be partitioned.

-> Gen [[a]]
The partitioned list.

Partitions a list into a list of sublists.

Each sublist in the result has a randomly-chosen length that is bounded by the given minimum and maximum length parameters, with the exception of the last sublist, which may be shorter than the minimum length.

Examples:

>>> generate (partitionList (0, 1) [1 .. 4])
[[], [1], [2], [], [3], [4]]

>>> generate (partitionList (1, 4) [1 .. 10])
[[1, 2], [3, 4, 5, 6], [7], [8, 9, 10]]

>>> generate (partitionList (4, 8) "Books are the liberated spirits of men.")
["Books ar", "e the l", "iberat", "ed spir", "its of", " men."]

Assuming the following definitions of checked minimum and maximum lengths:

>>> x' = max 0 x
>>> y' = max 1 (max y x')

This function satisfies the following properties:

forAll (partitionList (x, y) as) $ (== as) . mconcat

forAll (partitionList (x, y) as) $ all ((>= x') . length) . dropEnd 1

forAll (partitionList (x, y) as) $ all ((<= y') . length)

Selecting entries from maps

selectMapEntry :: forall k v. Ord k => Map k v -> Gen (Maybe ((k, v), Map k v))

Selects an entry at random from the given map.

Returns the selected entry and the remaining map with the entry removed.

Returns Nothing if (and only if) the given map is empty.

selectMapEntries :: forall k v. Ord k => Map k v -> Int -> Gen ([(k, v)], Map k v)

Selects up to a given number of entries at random from the given map.

Returns the selected entries and the remaining map with the entries removed.

Generating and shrinking natural numbers

chooseNatural :: (Natural, Natural) -> Gen Natural

shrinkNatural :: Natural -> [Natural]

Generating and shrinking non-empty lists

genNonEmpty :: Gen a -> Gen (NonEmpty a)

shrinkNonEmpty :: (a -> [a]) -> NonEmpty a -> [NonEmpty a]

Counterexamples

report :: (Show a, Testable prop) => a -> String -> prop -> Property

Adds a named variable to the counterexample output of a property.

On failure, uses pretty-printing to show the contents of the variable.

verify :: Testable t => Bool -> String -> t -> Property

Adds a named condition to a property.

On failure, reports the name of the condition that failed.

Pretty-printing

newtype Pretty a

A combinator that causes the output of show to be pretty-printed.

Constructors

Pretty

Fields

unPretty :: a

Combinators

newtype NotNull a

Constructors

NotNull

Fields

unNotNull :: a

Utilities

interleaveRoundRobin :: [[a]] -> [a]