Test.Utils.Env

Description

Documentation

withEnv :: MonadUnliftIO m => [(String, String)] -> m a -> m a

Runs an IO action with exactly the given environment variables. After the action finishes, the original environment variables are restored.

NOTE: The process environment is a global variable -- this function is not threadsafe.

withAddedEnv :: MonadUnliftIO m => [(String, String)] -> m a -> m a

Runs an IO action with the given environment variables set, in addition to the current environment variables. After the action finishes, the environment variables are restored back to the original state.

NOTE: The process environment is a global variable -- this function is not threadsafe.

clearEnv :: MonadIO m => m ()

Unsets all environment variables for this process.