Test.Utils.Ports

Description

Provides functions for checking if TCP ports are available to listen on. These can be used to start servers for testing when there may be multiple test suites running in parallel.

Includes code from nh2: https://stackoverflow.com/a/57022572

Documentation

randomUnusedTCPPorts :: Int -> IO [Int]

Get a list of random TCPv4 ports that currently do not have any servers listening on them. It may return less than the requested number of ports.

Note that this method of allocating ports is subject to race conditions. Production code should use better methods such as passing a listening socket to the child process.

findPort :: IO Int

Returen a single TCP port that was unused at the time this function was called.

isPortOpen :: SockAddr -> IO Bool

Checks whether connect() to a given TCPv4 SockAddr succeeds or returns eCONNREFUSED.

Rethrows connection exceptions in all other cases (e.g. when the host is unroutable).

simpleSockAddr :: (Word8, Word8, Word8, Word8) -> PortNumber -> SockAddr

Creates a SockAttr from host IP and port number.

Example: > simpleSockAddr (127,0,0,1) 8000