Test.Utils.Time

Description

Provides utility functions relating to testing with times and dates.

Documentation

data UniformTime

A wrapper for UTCTime whose Arbitrary instance spans a uniform range of dates and a mixture of time precisions.

genUniformTime :: Gen UTCTime

Generate UTCTime values over a uniform range of dates and a mixture of time precisions.

Dates will be generated in a range that's bounded by defaultLowerBound and defaultUpperBound.

genUniformTimeWithinRange :: Day -> Day -> Gen UTCTime

Generate UTCTime values over a uniform range of dates and a mixture of time precisions.

Dates will be generated in a range that's bounded by the given minimum and maximum Julian day arguments.

getUniformTime :: UniformTime -> UTCTime