Test.Utils.Trace

Description

Provides functions for setting up and capturing logging so that expectations about logging can be asserted in test scenarios.

Documentation

withLogging :: MonadUnliftIO m => ((Tracer m msg, m [msg]) -> m a) -> m a

Run an action with a logging Trace object, and a function to get all messages that have been traced.

withLogging' :: forall m mtr msg a. (MonadUnliftIO m, MonadIO mtr) => ((Tracer mtr msg, m [msg]) -> m a) -> m a

Same as withLogging, but with a different Tracer monad.

captureLogging :: MonadUnliftIO m => (Tracer m msg -> m a) -> m ([msg], a)

Run an action with a Trace, returning captured log messages along with the result of the action.

captureLogging' :: forall m mtr msg a. (MonadUnliftIO m, MonadIO mtr) => (Tracer mtr msg -> m a) -> m ([msg], a)

Same as captureLogging, but with a different Tracer monad.

traceSpec :: (HasCallStack, ToText msg) => SpecWith (Tracer IO msg) -> Spec

Provides a Tracer to the spec, which is silent, unless something goes wrong. In that case, it dumps all the traces it has collected to stdout.