Test.Utils.Paths

Description

Utility function for finding the package test data directory.

Documentation

getTestData :: Q Exp

A TH function to get the test data directory.

It combines the current source file location and cabal file to locate the package directory in such a way that works in both the stack/cabal package build and ghci.

For the Nix build, rather than baking in a path that starts with /build, it makes the test data path relative to the current directory.