Test.Utils.Time

Description

Provides utility functions relating to testing with times and dates.

Documentation

data UniformTime

A wrapper for UTCTime whose Arbitrary instance spans a uniform range of dates and a mixture of time precisions.

genUniformTime :: Gen UTCTime

Generate UTCTime values over a uniform range of dates and a mixture of time precisions.

getUniformTime :: UniformTime -> UTCTime